Remove UI status caching and reset
Trying to keep track of the UI status before an operation and resetting
it afterwards does not work well for us if these operations overlap.
Therefore we remove that feature for the time being.  This could
probably be solved by using counters to keep track of the number of
ongoing operations of a certain type.